requirements.txt
beautifulsoup4==4.8.0
discord.py==1.2.3
factorio-rcon-py==1.0.2
requests==2.22.0
SQLAlchemy==1.3.7
flask
psycopg2